Move most KnownRuntimeEffects into the sksl_rt_shader module.
The remaining KnownRuntimeEffects are assembled programmatically,
or are so simple that they wouldn't benefit from existing in a
module.

Bug: b/328051548
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/883098
Commit-Queue: John Stiles <johnstiles@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>

+++ src/effects/imagefilters/SkMorphologyImageFilter.cpp
@@ -25,7 +25,6 @@
 #include "src/core/SkWriteBuffer.h"
 
 #include <algorithm>
-#include <cstdint>
 #include <optional>
 #include <utility>
 
@@ -128,7 +127,7 @@ sk_sp<SkImageFilter> make_morphology(MorphType type,
 // room. The other tradeoff is that for R > kMaxLinearRadius, the sparse morphology kernel only
 // requires 2 samples to double the accumulated kernel size, but at the cost of another render
 // target.
-static constexpr int kMaxLinearRadius = 14; // KEEP IN SYNC W/ SkKnownRuntimeEffects.cpp VERSION
+static constexpr int kMaxLinearRadius = 14; // KEEP IN SYNC WITH CONSTANT IN `sksl_rt_shader.sksl`
 sk_sp<SkShader> make_linear_morphology(sk_sp<SkShader> input,
                                        MorphType type,
                                        MorphDirection direction,
@@ -142,7 +141,7 @@ sk_sp<SkShader> make_linear_morphology(sk_sp<SkShader> input,
     builder.child("child") = std::move(input);
     builder.uniform("offset") = direction == MorphDirection::kX ? SkV2{1.f, 0.f} : SkV2{0.f, 1.f};
     builder.uniform("flip") = (type == MorphType::kDilate) ? 1.f : -1.f;
-    builder.uniform("radius") = (int32_t) radius;
+    builder.uniform("radius") = (float)radius;
 
     return builder.makeShader();
 }
